Add files via upload
This project is a creative and visually impressive use of HTML and CSS to create the Google Assistant GIF or logo. By using HTML and CSS to replicate the iconic Google Assistant graphic.

The website likely features the GIF or logo in a prominent location on the page, and the use of HTML and CSS allows to create an interactive and engaging user experience. This include animated elements that make the logo or GIF come to life on the page.

The use of HTML and CSS also allows you to manipulate the visual design of the logo or GIF, experimenting with color schemes, gradients, and other visual elements to create a unique and eye-catching display. By using these tools creatively, we can create a striking and memorable representation of the Google Assistant logo or GIF.

Overall, This project is a great demonstration of the versatility of HTML and CSS, and a creative use of these web development tools to create an interactive and visually appealing display of the Google Assistant logo or GIF. It is sure to impress and engage visitors to your website.